+++ tests/algorithms/multiway/test_fmgm.py
@@ -8,28 +8,6 @@ import graph_matching_tools.algorithms.kernels.gaussian as kern
 
 
 class TestFMGM(unittest.TestCase):
-    def test_create_pairwise_gradient(self):
-        inc1 = np.array([[1, 0], [1, 1], [0, 1]])
-        inc2 = np.array([[0, 1], [1, 1], [1, 1]])
-        knode = np.array([[0.0, 1.0, 2.0], [0.2, 1.0, 2.0], [0.9, 0.8, 0.2]])
-        kedge = np.array([[0.0, 1.0], [0.2, 2.0]])
-
-        grad = fmgm.create_pairwise_gradient(inc1, inc2, knode, kedge)
-        res = grad(np.ones((3, 3)))
-        truth = np.array([[4.0, 6.0, 8.0], [12.4, 14.4, 16.4], [9.8, 10.0, 8.8]])
-        self.assertTrue(np.linalg.norm(res - truth) < 1e-2)
-
-    def test_rrwm(self):
-        inc1 = np.array([[1, 0], [1, 1], [0, 1]])
-        inc2 = np.array([[0, 1], [1, 1], [1, 1]])
-        knode = np.array([[0.0, 1.0, 2.0], [0.2, 1.0, 2.0], [0.9, 0.8, 0.2]])
-        kedge = np.array([[0.0, 1.0], [0.2, 2.0]])
-
-        grad = fmgm.create_pairwise_gradient(inc1, inc2, knode, kedge)
-        perm = fmgm.rrwm(grad, (3, 3))
-        truth = np.array([[0.0, 0.0, 1.0], [0.0, 1.0, 0.0], [1.0, 0.0, 0.0]])
-        self.assertTrue(np.linalg.norm(perm - truth) < 1e-2)
-
     def test_factorized_multigraph_matching(self):
         node_kernel = kern.create_gaussian_node_kernel(2.0, "weight")